Safari-609.1.11.1/LayoutTests/http/tests/websocket/tests/hybi/non-document-mixed-content-blocked-https-with-embedded-http-with-embedded-https.https.html

<!DOCTYPE HTML>
<html>
<head>
<script>

if (window.testRunner) {
    testRunner.waitUntilDone();
    testRunner.dumpAsText();
}
if (window.internals)
    internals.settings.setAllowDisplayOfInsecureContent(true);

</script>
</head>
<body>
<iframe src="http://localhost:8000/websocket/tests/hybi/non-document-mixed-content-blocked-http-with-embedded-https.html"></iframe>
<p>Test topFrame https with embedded http with embedded https blocks worker trying to connect with insecure webSocket.
</p>
<p></p>
<pre id=log>
</pre>
</body>
</html>
